Fix Klarna integration (CS1)
Existing implementation didn't include environment specific base URL applicable
to Klarna client operations. This commit adds the ability to use the `SANDBOX`
base url and integration tests for create and update session scenarios.

It also fixes mapping issues in request/response classes. In fields related to
amounts, Integers were replaced by Long types.

+++ src/main/java/com/checkout/apm/klarna/CreditSessionRequest.java
@@ -19,10 +19,10 @@ public final class CreditSessionRequest {
 
     private String locale;
 
-    private Integer amount;
+    private Long amount;
 
     @SerializedName("tax_amount")
-    private Integer taxAmount;
+    private Long taxAmount;
 
     private List<KlarnaProduct> products;
